Seperated out code of manager and made shell-like interactivity
The manager has now moved into the 'man' folder as it is now a little more organized.
I tried to make it cleaner and everything safer. It now supports an interactive
shell-like session with the '-i' flag. Right now the return values that the
manager sends back are not the best, but I will most likely update those in the
future. For now, I think this is a large step forward for the manager.